+++ src/components/property/review/new/StarRating.tsx
@@ -1,39 +1,25 @@
 "use client";
 
-import { useState } from "react";
+import { useStarRating } from "@/hooks/property/useStarRating";
 
 const StarRating = () => {
-  const [rating, setRating] = useState(0);
-
-  const handleClick = (index: number, isHalf: boolean) => {
-    const newRating = isHalf ? index + 0.5 : index + 1;
-    setRating(newRating);
-  };
-
-  const renderStarIcon = (index: number) => {
-    if (rating >= index + 1) return "/icons/star-filled.svg";
-    if (rating >= index + 0.5) return "/icons/star-half.svg";
-    return "/icons/star-unfilled.svg";
-  };
+  const { rating, handleClick, renderStarIcon } = useStarRating();
 
   return (
     <div className="flex items-center justify-between">
       <div className="flex gap-[4px]">
         {[0, 1, 2, 3, 4].map((i) => (
           <div key={i} className="relative h-8 w-8">
-            {/* 왼쪽 (0.5점) */}
             <button
               type="button"
               className="absolute left-0 top-0 z-10 h-full w-1/2"
               onClick={() => handleClick(i, true)}
             />
-            {/* 오른쪽 (1점) */}
             <button
               type="button"
               className="absolute right-0 top-0 z-10 h-full w-1/2"
               onClick={() => handleClick(i, false)}
             />
-            {/* 별 아이콘 */}
             <img
               src={renderStarIcon(i)}
               alt="star"
